+++ CHANGELOG.md
@@ -74,6 +74,7 @@
     - Added info message at the top of the contact form
     - Can be opened from other components using props, which allow a different addressee (to) to be used from the one set in config.json
 - Print: Parameter additionalLayer to add layer (e.g. coordinate grid) during print.
+- Refactored Modeler3D to work with MP 3.0
 
 ### Changed
 - Eslint: no-undef eslint rule is switched on.